Add build options --resolv and --hosts
Add flags --resolv and --hosts for `apptainer build` to determine the
path to be mounted on the container `/etc/resolv.conf` and `/etc/hosts`
when building the image.

This allows unprivileged users to build containers without relying on
the top-level directories.

Changed function interfaces:

In util.go: createStageFile: add `destSubpath` as the second input
argument.
(This function is currently used to prepare the temporary files
to mount to `/etc/resolv.conf` and `/etc/hosts`; the impact to other
parts of code should be minimal)

+++ internal/pkg/build/build.go
@@ -408,17 +408,23 @@ func (b *Build) Full(ctx context.Context) error {
 		}
 
 		// create stage file for /etc/resolv.conf and /etc/hosts
-		sessionResolv, err := createStageFile("/etc/resolv.conf", stage.b, "Name resolution could fail")
-		if err != nil {
-			return err
-		} else if sessionResolv != "" {
-			defer os.Remove(sessionResolv)
+		sessionResolv := ""
+		if stage.b.Opts.ResolvConfPath != "" {
+			sessionResolv, err = createStageFile(stage.b.Opts.ResolvConfPath, "/etc/resolv.conf", stage.b, "Name resolution could fail")
+			if err != nil {
+				return err
+			} else if sessionResolv != "" {
+				defer os.Remove(sessionResolv)
+			}
 		}
-		sessionHosts, err := createStageFile("/etc/hosts", stage.b, "Host resolution could fail")
-		if err != nil {
-			return err
-		} else if sessionHosts != "" {
-			defer os.Remove(sessionHosts)
+		sessionHosts := ""
+		if stage.b.Opts.HostsPath != "" {
+			sessionHosts, err = createStageFile(stage.b.Opts.HostsPath, "/etc/hosts", stage.b, "Host resolution could fail")
+			if err != nil {
+				return err
+			} else if sessionHosts != "" {
+				defer os.Remove(sessionHosts)
+			}
 		}
 
 		if stage.b.Recipe.BuildData.Post.Script != "" {
